Added basic functionality to components
Did all but one of the components that I wanted for this website. It has very basic functionality and will need to be adjusted later on. Hard to say what is missing without a running website but most of it should be good to go as it pulls exercise information from the wger workout manager api.